+++ frontend/src/tests/lib/utils/sns-neuron.utils.spec.ts
@@ -45,7 +45,6 @@ import {
   isSnsNeuron,
   isUserHotkey,
   isVesting,
-  maturityPercentageToE8s,
   minNeuronSplittable,
   needsRefresh,
   neuronAge,
@@ -2429,26 +2428,6 @@ describe("sns-neuron utils", () => {
           vesting_period_seconds: [oneWeek],
         })
       ).toEqual(0n);
-    });
-  });
-
-  describe("maturityPercentageToE8s", () => {
-    it("calculates percents ", () => {
-      expect(
-        maturityPercentageToE8s({
-          total: 100,
-          percentage: 50,
-        })
-      ).toEqual(5_000_000_000n);
-    });
-
-    it("handles more than 8 decimals results", () => {
-      expect(
-        maturityPercentageToE8s({
-          total: 1.00001,
-          percentage: 1,
-        })
-      ).toEqual(1_000_010n);
     });
   });
 });
